examples/zillow_api_challenge/requirements.txt
Django==2.2.20
django-filter==2.4.0
djangorestframework==3.12.2
pytz==2020.5
